refactor: extract useDataPersistence hook
